Keep the prescan counters out of the player-facing stats
getProfileStats() otherwise reports only what is in the profile, so two
fields describing the engine sat oddly in it and invited multithreading
to be blamed for unrelated trouble. They are what stops the flood spec
from silently testing nothing, so they stay - behind MUDLET_TEST_MODE,
as waitForEvent() and pumpEvents() already do.

Also drops the C type long, which CodeFactor flags.

+++ src/TriggerMatchPool.cpp
@@ -120,7 +120,7 @@ void TriggerMatchPool::runChunks(int slot)
 void TriggerMatchPool::workerLoop(int slot, uint64_t startEpoch)
 {
     uint64_t seen = startEpoch;
-    long long idle = 0;
+    qint64 idle = 0;
     for (;;) {
         const uint64_t epoch = mEpoch.load(std::memory_order_acquire);
         if (epoch != seen) {
